Python Password-Generator Thing
Sorry for all of the comments, it helps me learn to a deeper level what things do when I explain them myself.

This is just a basic password generator that can run in terminal/cmd/powershell

This is a little something that I added to another program which I'll probably throw on here later.